pyproject.toml
[tool.poetry]
name = "aithena-memory"
version = "0.1.0"
description = ""
authors = ["Elder Millenial <eldermillenial@protonmail.com>"]
readme = "README.md"
packages = [{ include = "aithena", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.9.1"
psycopg = { extras = ["binary", "pool"], version = "^3.2.1" }
python-dotenv = "^1.0.1"
pgvector = "^0.3.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
